--- sim/core/dbc/spell.go
@@ -1,1 +1,0 @@
-package dbc
